+++ flowman-tools/src/main/scala/com/dimajix/flowman/tools/exec/job/JobCommand.scala
@@ -32,7 +32,6 @@ class JobCommand extends NestedCommand {
     @SubCommands(Array(
         new SubCommand(name="list",impl=classOf[ListCommand]),
         new SubCommand(name="create",impl=classOf[CreateCommand]),
-        new SubCommand(name="migrate",impl=classOf[MigrateCommand]),
         new SubCommand(name="build",impl=classOf[BuildCommand]),
         new SubCommand(name="verify",impl=classOf[VerifyCommand]),
         new SubCommand(name="truncate",impl=classOf[TruncateCommand]),