+++ flowman-tools/src/main/scala/com/dimajix/flowman/tools/exec/target/PhaseCommand.scala
@@ -77,7 +77,6 @@ class PhaseCommand(phase:Phase) extends ActionCommand {
 
 
 class CreateCommand extends PhaseCommand(Phase.CREATE)
-class MigrateCommand extends PhaseCommand(Phase.MIGRATE)
 class BuildCommand extends PhaseCommand(Phase.BUILD)
 class VerifyCommand extends PhaseCommand(Phase.VERIFY)
 class TruncateCommand extends PhaseCommand(Phase.TRUNCATE)